Cleanup unused code
