+++ Sources/UI/Resources/Snabble.xcassets/discount-badge.imageset/discount-badge.pdf
@@ -1,0 +1,184 @@
+%PDF-1.7
+
+1 0 obj
+  << /Type /XObject
+     /Length 2 0 R
+     /Group << /Type /Group
+               /S /Transparency
+            >>
+     /Subtype /Form
+     /Resources << >>
+     /BBox [ 0.000000 0.000000 16.000000 16.000000 ]
+  >>
+stream
+/DeviceRGB CS
+/DeviceRGB cs
+q
+1.000000 0.000000 -0.000000 1.000000 0.000000 0.000000 cm
+1.000000 0.250980 0.505882 scn
+0.000000 8.000000 m
+0.000000 12.418278 3.581722 16.000000 8.000000 16.000000 c
+8.000000 16.000000 l
+12.418278 16.000000 16.000000 12.418278 16.000000 8.000000 c
+16.000000 8.000000 l
+16.000000 3.581722 12.418278 0.000000 8.000000 0.000000 c
+8.000000 0.000000 l
+3.581722 0.000000 0.000000 3.581722 0.000000 8.000000 c
+0.000000 8.000000 l
+h
+f
+n
+Q
+q
+1.000000 0.000000 -0.000000 1.000000 3.333374 3.333344 cm
+0.984314 0.984314 1.000000 scn
+2.041667 9.333328 m
+0.915833 9.333328 0.000000 8.417495 0.000000 7.291661 c
+0.000000 6.165828 0.915833 5.249995 2.041667 5.249995 c
+3.167500 5.249995 4.083333 6.165828 4.083333 7.291661 c
+4.083333 8.417495 3.167500 9.333328 2.041667 9.333328 c
+h
+2.041667 6.416661 m
+1.557500 6.416661 1.166667 6.807495 1.166667 7.291661 c
+1.166667 7.775828 1.557500 8.166661 2.041667 8.166661 c
+2.525834 8.166661 2.916667 7.775828 2.916667 7.291661 c
+2.916667 6.807495 2.525834 6.416661 2.041667 6.416661 c
+h
+7.291667 4.083328 m
+6.165834 4.083328 5.250000 3.167494 5.250000 2.041661 c
+5.250000 0.915828 6.165834 -0.000006 7.291667 -0.000006 c
+8.417500 -0.000006 9.333334 0.915828 9.333334 2.041661 c
+9.333334 3.167494 8.417500 4.083328 7.291667 4.083328 c
+h
+7.291667 1.166661 m
+6.807500 1.166661 6.416667 1.557494 6.416667 2.041661 c
+6.416667 2.525828 6.807500 2.916661 7.291667 2.916661 c
+7.775834 2.916661 8.166667 2.525828 8.166667 2.041661 c
+8.166667 1.557494 7.775834 1.166661 7.291667 1.166661 c
+h
+0.822500 -0.000006 m
+0.000000 0.822495 l
+8.510834 9.333328 l
+9.333334 8.510828 l
+0.822500 -0.000006 l
+h
+f
+n
+Q
+
+endstream
+endobj
+
+2 0 obj
+  1651
+endobj
+
+3 0 obj
+  << /Type /XObject
+     /Length 4 0 R
+     /Group << /Type /Group
+               /S /Transparency
+            >>
+     /Subtype /Form
+     /Resources << >>
+     /BBox [ 0.000000 0.000000 16.000000 16.000000 ]
+  >>
+stream
+/DeviceRGB CS
+/DeviceRGB cs
+q
+1.000000 0.000000 -0.000000 1.000000 0.000000 0.000000 cm
+0.000000 0.000000 0.000000 scn
+0.000000 8.000000 m
+0.000000 12.418278 3.581722 16.000000 8.000000 16.000000 c
+8.000000 16.000000 l
+12.418278 16.000000 16.000000 12.418278 16.000000 8.000000 c
+16.000000 8.000000 l
+16.000000 3.581722 12.418278 0.000000 8.000000 0.000000 c
+8.000000 0.000000 l
+3.581722 0.000000 0.000000 3.581722 0.000000 8.000000 c
+0.000000 8.000000 l
+h
+f
+n
+Q
+
+endstream
+endobj
+
+4 0 obj
+  463
+endobj
+
+5 0 obj
+  << /XObject << /X1 1 0 R >>
+     /ExtGState << /E1 << /SMask << /Type /Mask
+                                    /G 3 0 R
+                                    /S /Alpha
+                                 >>
+                          /Type /ExtGState
+                       >> >>
+  >>
+endobj
+
+6 0 obj
+  << /Length 7 0 R >>
+stream
+/DeviceRGB CS
+/DeviceRGB cs
+q
+/E1 gs
+/X1 Do
+Q
+
+endstream
+endobj
+
+7 0 obj
+  46
+endobj
+
+8 0 obj
+  << /Annots []
+     /Type /Page
+     /MediaBox [ 0.000000 0.000000 16.000000 16.000000 ]
+     /Resources 5 0 R
+     /Contents 6 0 R
+     /Parent 9 0 R
+  >>
+endobj
+
+9 0 obj
+  << /Kids [ 8 0 R ]
+     /Count 1
+     /Type /Pages
+  >>
+endobj
+
+10 0 obj
+  << /Pages 9 0 R
+     /Type /Catalog
+  >>
+endobj
+
+xref
+0 11
+0000000000 65535 f
+0000000010 00000 n
+0000001909 00000 n
+0000001932 00000 n
+0000002643 00000 n
+0000002665 00000 n
+0000002963 00000 n
+0000003065 00000 n
+0000003086 00000 n
+0000003259 00000 n
+0000003333 00000 n
+trailer
+<< /ID [ (some) (id) ]
+   /Root 10 0 R
+   /Size 11
+>>
+startxref
+3393
+%%EOF
